requirements.txt
openai
google-generativeai
pyyaml
python-dotenv
requests
pandas
